pyproject.toml
[project]
name = "legal-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.8",
    "langgraph-api>=0.4.31",
    "langgraph-cli>=0.4.2",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.4",
    "chromadb>=0.5.5",
    "python-dotenv>=1.0.1",
]
